refactor: extract Samples logic into NavBar/Samples component
Moves sample store state, load handlers, and dropdown rendering out of
Navbar into a dedicated Samples component. Shared UI primitives extracted
to NavBar/components.tsx. SampleMenuItems exported for reuse in mobile menu.
