+++ src/components/NavBar/components.tsx
@@ -1,0 +1,144 @@
+import { useState } from "react";
+import { Link } from "react-router-dom";
+
+export interface DropdownProps {
+  children: React.ReactNode;
+  overlay: React.ReactNode;
+  trigger: string[];
+  className?: string;
+}
+
+export const Dropdown = ({
+  children,
+  overlay,
+  trigger,
+  className = "",
+}: DropdownProps) => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const handleClick = () => {
+    if (trigger.includes("click")) {
+      setIsOpen(!isOpen);
+    }
+  };
+
+  const handleMouseEnter = () => {
+    if (trigger.includes("hover")) {
+      setIsOpen(true);
+    }
+  };
+
+  const handleMouseLeave = () => {
+    if (trigger.includes("hover")) {
+      setIsOpen(false);
+    }
+  };
+
+  return (
+    <div
+      className={`relative ${className}`}
+      onMouseEnter={handleMouseEnter}
+      onMouseLeave={handleMouseLeave}
+    >
+      <div onClick={handleClick}>{children}</div>
+      {isOpen && (
+        <>
+          <div
+            className="fixed inset-0 z-10"
+            onClick={() => setIsOpen(false)}
+          />
+          <div className="absolute top-full left-0 z-20 mt-1 min-w-48 bg-white dark:bg-gray-800 rounded-md shadow-lg border border-gray-200 dark:border-gray-700">
+            {overlay}
+          </div>
+        </>
+      )}
+    </div>
+  );
+};
+
+export const Menu = ({
+  children,
+  className = "",
+}: {
+  children: React.ReactNode;
+  className?: string;
+}) => <div className={`py-1 ${className}`}>{children}</div>;
+
+export const MenuItem = ({
+  children,
+  onClick,
+  to,
+  href,
+  className = "",
+}: {
+  children: React.ReactNode;
+  onClick?: () => void;
+  to?: string;
+  href?: string;
+  className?: string;
+}) => {
+  const baseClasses = `px-4 py-2 text-sm text-gray-700 dark:text-gray-200 hover:bg-gray-100 dark:hover:bg-gray-700 cursor-pointer flex items-center space-x-2 ${className}`;
+
+  if (to) {
+    return (
+      <Link to={to} className={baseClasses} onClick={onClick}>
+        {children}
+      </Link>
+    );
+  }
+
+  if (href) {
+    return (
+      <a
+        href={href}
+        target="_blank"
+        rel="noopener noreferrer"
+        className={baseClasses}
+        onClick={onClick}
+      >
+        {children}
+      </a>
+    );
+  }
+
+  return (
+    <button
+      type="button"
+      className={`w-full text-left bg-transparent border-none ${baseClasses}`}
+      onClick={onClick}
+    >
+      {children}
+    </button>
+  );
+};
+
+export const MenuItemGroup = ({
+  title,
+  children,
+  className = "",
+}: {
+  title: string;
+  children: React.ReactNode;
+  className?: string;
+}) => (
+  <div className={className}>
+    <div className="px-4 py-2 text-xs font-semibold text-gray-500 dark:text-gray-400 uppercase tracking-wide border-b border-gray-200 dark:border-gray-700">
+      {title}
+    </div>
+    {children}
+  </div>
+);
+
+export const Button = ({
+  children,
+  onClick,
+  className = "",
+}: {
+  children: React.ReactNode;
+  onClick?: () => void;
+  className?: string;
+}) => (
+  <button className={`flex items-center ${className}`} onClick={onClick}>
+    {children}
+  </button>
+);
